requirements.txt
porespy==1.0
pyevtk==1.1.1
numpy-stl==2.10.0